refactor: move language service logic
