+++ packages/ls-core/src/index.ts
@@ -1,2 +1,8 @@
 export { tsMdLanguagePlugin as createTsMdPlugin } from './plugin.js';
 export type { TsMdVirtualFile } from './virtual-file.js';
+export {
+  createTsMdLanguageService,
+  collectDiagnostics,
+  type TsMdDiagnostic,
+  type TsMdDiagnosticsResult,
+} from './service.js';
