+++ packages/ls-core/tsup.config.ts
@@ -6,4 +6,5 @@ export default defineConfig({
   dts: true,
   clean: true,
   target: 'node18',
+  external: ['typescript'],
 });
